Stage 4 - change from using a HashMap to a 2D array for JButtons
While a HashMap<String, ArrayList<JButton>> worked for the prior stage
in playing moves, it does not lend itself to effectively checking the
playing field for four in a row. To fix this, we switch to a 2D array
of JButton's (JButton[][]) so we can just use indexes to check each of
the directions (horizontal, vertical and diagonal).

ConnectFour.java
- Define constants for number of rows and columns, as well as button
  size. Window size is calculated from these.
- Change addComponents() to use an array for storing the buttons.
- Change the button.addActionListener to a lambda which calls the
  appropriate method in the controller.
- Add a panel for holding the reset button needed for the stage.

ConnectFourController.java
- No longer implements ActionListener; we use lambdas to call methods
  in this class to handle the actions we want.
- Change setButtonMap() to accept a JButton array instead of the HashMap
- Move the code in playCell() to ConnectFourModel.java and call that
  method from here.

ConnectFourModel.java
- New class to contain all the code/logic for playing the game and
  detecting the winner